requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge_Score
py7zr
nltk
tqdm
PyYAML
mypy-boto3-s3
ensure
python-box==6.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2

-e .